Split up federation code into multiple files
