+++ Dockerfile
@@ -5,6 +5,6 @@ ENV TARGET_BUCKET=replace_me
 COPY requirements.txt .
 RUN pip install -r requirements.txt
 
-COPY athena_federator.py gathena.py credentials.json token.pickle ./
+COPY athena/ gathena.py credentials.json token.pickle ./
 
 CMD [ "gathena.lambda_handler" ]
